ability to change user roles and protect the delete func

+++ convex/files.ts
@@ -32,7 +32,7 @@ async function hasAccessToOrg(ctx: QueryCtx|MutationCtx, orgId: string) {
     if(!user){
         return null;
     }
-    const hasAccess = user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    const hasAccess = user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId);
 
     if(!hasAccess){
         return null;
@@ -115,6 +115,11 @@ export const deleteFile = mutation({
         const access = await hasAccessToFile(ctx, args.fileId);
         if(!access){
             throw new ConvexError("you have no access to file");
+        }
+
+        const isAdmin = access.user.orgIds.find((org) => org.orgId === access.file.orgId)?.role==="admin";
+        if(!isAdmin){
+            throw new ConvexError("you have no admin access to delete");
         }
         await ctx.db.delete(args.fileId);
     },
